+++ core/CHANGELOG.md
@@ -4,10 +4,13 @@
 
 - Hide `prost::Error` from public API in `FromEnvelopeError::InvalidPeerRecord` and `signed_envelope::DecodingError`. See [PR 3058].
 
+- Move `Executor` to `libp2p-swarm`. See [PR 3097].
+
 - Improve error messages in case keys cannot be decoded because of missing feature flags. See [PR 2972].
 
 [PR 3031]: https://github.com/libp2p/rust-libp2p/pull/3031
 [PR 3058]: https://github.com/libp2p/rust-libp2p/pull/3058
+[PR 3097]: https://github.com/libp2p/rust-libp2p/pull/3097
 [PR 2972]: https://github.com/libp2p/rust-libp2p/pull/2972
 
 # 0.37.0